Remove requirement for inheritance for cell handlers
