+++ src/cellHandler/default.ts
@@ -2,11 +2,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-import { CellHandler, CellHandlerAttachParameters } from "./base";
+import { BaseCellHandler } from "./base";
 import { render, html } from "lit-html";
 import { StarboardTextEditor } from "../components/textEditor";
 import { Cell } from "../runtime/types";
-import { Runtime } from "../runtime";
+import { Runtime, CellHandlerAttachParameters } from "../runtime";
 
 export const DEFAULT_CELL_TYPE_DEFINITION = {
     name: "Unknown",
@@ -17,7 +17,7 @@ export const DEFAULT_CELL_TYPE_DEFINITION = {
 /**
  * The cell handler that gets used when there is an unknown cell type
  */
-export class DefaultCellHandler extends CellHandler {
+export class DefaultCellHandler extends BaseCellHandler {
     constructor(cell: Cell, runtime: Runtime) {
         super(cell, runtime);
     }
